Moved SampleGroupProcessors
