+++ src/smartpeak/source/core/SampleGroupProcessors/MergeInjections.cpp
@@ -723,73 +723,4 @@ namespace SmartPeak
     feature_map.erase(feature_map.begin());
   }
 
-
-  ParameterSet LoadFeaturesSampleGroup::getParameterSchema() const
-  {
-    return ParameterSet();
-  }
-
-  void LoadFeaturesSampleGroup::getFilenames(Filenames& filenames) const
-  {
-    filenames.addFileName("featureXMLSampleGroup_i", "${FEATURES_INPUT_PATH}/${INPUT_GROUP_NAME}.featureXML");
-  };
-
-  void LoadFeaturesSampleGroup::process(SampleGroupHandler& sampleGroupHandler_IO,
-    const SequenceHandler& sequenceHandler_I,
-    const ParameterSet& params_I,
-    Filenames& filenames_I
-  ) const
-  {
-    LOGD << "START LoadFeaturesSampleGroup";
-    getFilenames(filenames_I);
-
-    if (!InputDataValidation::prepareToLoad(filenames_I, "featureXMLSampleGroup_i"))
-    {
-      throw std::invalid_argument("Failed to load input file");
-    }
-
-    try {
-      OpenMS::FeatureXMLFile featurexml;
-      featurexml.load(filenames_I.getFullPath("featureXMLSampleGroup_i").generic_string(), sampleGroupHandler_IO.getFeatureMap());
-    }
-    catch (const std::exception& e) {
-      LOGE << e.what();
-      sampleGroupHandler_IO.getFeatureMap().clear();
-      LOGE << "feature map clear";
-      throw e;
-    }
-
-    LOGD << "END LoadFeaturesSampleGroup";
-  }
-
-  ParameterSet StoreFeaturesSampleGroup::getParameterSchema() const
-  {
-    return ParameterSet();
-  }
-
-  void StoreFeaturesSampleGroup::getFilenames(Filenames& filenames) const
-  {
-    filenames.addFileName("featureXMLSampleGroup_o", "${FEATURES_OUTPUT_PATH}/${OUTPUT_GROUP_NAME}.featureXML");
-  };
-
-  void StoreFeaturesSampleGroup::process(SampleGroupHandler& sampleGroupHandler_IO,
-    const SequenceHandler& sequenceHandler_I,
-    const ParameterSet& params_I,
-    Filenames& filenames_I
-  ) const
-  {
-    LOGD << "START storeFeaturesSampleGroup";
-    getFilenames(filenames_I);
-
-    if (!InputDataValidation::prepareToStore(filenames_I, "featureXMLSampleGroup_o"))
-    {
-      throw std::invalid_argument("Failed to store output file");
-    }
-
-    // Store outfile as featureXML
-    OpenMS::FeatureXMLFile featurexml;
-    featurexml.store(filenames_I.getFullPath("featureXMLSampleGroup_o").generic_string(), sampleGroupHandler_IO.getFeatureMap());
-
-    LOGD << "END storeFeaturesSampleGroup";
-  }
 }
